Refactor Preview Context into new File Actions Context
We'll be reusing both the file preview functionality and the other actions
(delete, rename, download, new version, view versions) in multiple places in
the app, so we create a new `components/files/FileActions/FileActionsContext`
to manage dialog state for all the actions, as well as the pagination, loading,
and error state for the previews.

During this process, it became clear that some queries and mutations that had
been included in the `ProjectFiles.graphql` and labeled as `Project`-specific
should really be agnostic. These were moved and renamed, and typings
regenerated.