Tighten up typings for File Actions
The File Actions Context had taken just `File` and `Directory` as options for
the types upon with it could perform actions. Since most of our queried data
comes in the form of only partial implementations of these types, and since we
also now need to be able to pass in `FileVersion` implementations, this was
starting to create a lot of tedious and unsafe type assertions.

This updates the File Actions Context, Project Files, File Versions, and File
Version Item components to use more specific typings. Going forward, any time
a new scene or component consumes the File Actions Context, we may need to
add to the typings it accepts.

We also:

- Move the `FileNodeInfo` fragment from the Projects scene to the
`components/files` folder, since it is starting to be consumed by other scenes.
- Rename `FileActionItem` type in File Actions Context to `FilesActionItem` to
  make it clear that this is not about items of type `File` but about the
  "Files" feature
- Split exported types from `ProjectFilesList` into "directory" and "file" types
  to make it easier to consume these separately later, if necessary