Refactor preview file retrieval logic
We were repeating a lot of the same logic for fetching a file from the
`downloadUrl` passed in to the various Previewer components. This was done
initially because some Previewers could retrieve the file themselves from the
`downloadUrl`. In the interests of uniformity, those components have been
switched to accepting the file, which we now retrieve in the same way for every
Previewer at the `FilePreview` component level and pass down. This allows us
to DRY up our code while still leaving each individual component in control of
the actual file parsing/reading/presentation. We also delete our now-unused
`useRetrieveFile` hook.