Throw out previous logic for calculating merged rows/columns and write the correct version.
For some reason I thought it made sense to calculate merges by reading each row and column through in order, checking whether it had a merge in the array
provided by `SheetJS`, and then trying to modify all subsequent rows and columns
affected by that merge. This was brain-breakingly complex and required many
loops over the same data when multiple overlapping row merges existed.

Instead, this does the sensible thing, and first creates an array of all cells
that ought to be marked as spanned, then reads the data through one time,
returning `spanned` if spanned and the actual data if not.