Update `material-table` to 1.68.0
The bug keeping us at 1.60.0 has been resolved, but we need to still fix the
version at 1.68.0 instead of ^1.69.0 (the latest) because 1.69.0 uses a new
version of another library that throws a linter warning. There is apparently a
fix on the way for this soon.

We also update the patch to fix an issue with `aria-label` attributes being
incorrectly passed as props to the `MTableEditCell` component. I have opened
a PR to fix this, but I don't know how long their lead time is on getting to
those.