Remove faking focus on EnumField
This was just thrown in randomly to try to help sync FF focus state with browser state.
But it doesn't do that it just shows as focused even when the browser is unaware.

It also caused an infinite render loop with <FormControl /> which would try (and fail)
to set focus to false when disabled while rendering. While I don't agree that the MUI
code is the best, I can agree that we shouldn't show focused while field is disabled.
If we wanted to keep this code, we could change it to `active && !disabled`.
I started with this change but then realized the above paragraph.