Pass client timezone to server via cookie
This isn't perfect but it handles the majority use case of repeat visits in the same timezone.
For first visits or repeat visits where a timezone has changed the server could get the
formatting wrong but it would then be shortly corrected once client hydrates.

Definitely open to better solutions :/