Manage caching of API errors (for SSR) manually as Apollo does not store them
This allows errors retrieved during SSR to propagate to the client,
so the client uses the error the server received instead of executing
the query again.

Currently only the server is configured to write to this cache. In the future,
we could look at having the client write as well. This would allow not found errors
to be kept, but that also leads to stale cache problems as well.