Revert cd07db2f
hasMore is not used to determine if our cached list is complete.
It's used to communicate if there is a next page for our given "cursor".

So even if we have completed all 5 pages and we start at page 1, this should
go back to true because there is a page 2.