Remove itemFragment from addItemToList
We do not need to writeFragment() as the result of the mutation is already
written to the cache. We just need to grab a reference to it instead, with
Apollo has a helper for.

I'm thinking that this started from copying the documentation examples,
and then me blinding accepting it's necessity.