Improve selecting type object's list field.
For this apollo cache operation all we need from the object is the id and typename to
identity it. Whether or not it actually has the list field being referenced is irrelevant,
since the list field could be in the apollo cache from another operation.
Now we just enforce that the field key exists on the object type in the schema.

For example,
```
listId: [{ __typename: 'TranslationProject', id: "adsf" }, 'team']
```
This is now valid because `team` exists on the `TranslationProject` type in the gql schema,
even though it doesn't exist on the actual object given.