Fix modifyList (caching) to work with optimistic responses
Each optimistic response adds a Layer onto the Root entity store.
This allows the layers to be popped off or replaced with server response.
When accessing cache inside of an optimistic response the store is this
Layer instead of the Root store. This layer doesn't have any of the "main"
cache data, so the list variations could not be found.

Instead of forcing into the store's data, we call toObject() now which
returns its data with the parent layers merged in.
So now we can find all list variations outside of this optimistic response,
and modify this layer to have that list variation with the modified data.