Move authentication routes into <Root>
The main reason for this is to stop react-router from warning about
an unmatched route from the useRoutes() call.

I still kinda would like the authentication to be handled/grouped
outside this component. BUT, this does put all the top level routes
in one spot, and this file is much smaller now with the other refactors.